#42 truck

+++ app/models/truck.rb
@@ -15,7 +15,7 @@ class Truck < ActiveRecord::Base
   validates :truck_no, :numericality => true, :allow_blank => true
   validates :color, :format => { :with => ValidationValues.alpha, :message => message  }
   validates :license_expires, :format => { :with => ValidationValues.date, :message => message }, :allow_blank => true
-  validates :total_kilometres, :format => { :with => ValidationValues.double, :message => message }
+  validates :total_kilometres, :format => { :with => ValidationValues.integer, :message => message }
   validates :make, :model, :format => { :with => ValidationValues.alpha_numeric, :message => message }
   validates :vin_number, :format => { :with => ValidationValues.vin, :message => message }, :length => 17..17, :uniqueness => true
   validates :current_location, :format => { :with => ValidationValues.street, :message => message }
